meshed_prototype/watchover_scripting/pyproject.toml
[project]
name = "vncrediturst-watchtower"
version = "0.1.0"
description = "The watchtower component for VnCrediTrust"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "aiofiles>=25.1.0",
  "click>=8.3.0",
  "dataclass-settings>=0.7.0",
  "minify-html>=0.18.1",
  "msgspec>=0.19.0",
  "pydantic>=2.12.4",
  "pydantic-settings>=2.12.0",
  "pydoll-python>=2.12.1",
  "python-slugify[unidecode]>=8.0.4",
  "rich>=14.2.0",
  "structlog>=25.5.0",
  "trafilatura>=2.0.0",
  "types-lxml>=2025.8.25",
  "uvloop>=0.22.1",
]

[project.scripts]
watchtower = "app.__main__:run_cli"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling", "setuptools"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
dev-mode-dirs = ["src", "."]
ignore-vcs = true
sources = ["src"]

[tool.hatch.build.targets.sdist]
exclude = []
include = ["src/*", "tests/*", "public/*", "LICENSE"]
skip-excluded-dirs = false

[tool.hatch.build.targets.wheel]
packages = ["app"]
